Skal kunne velge stønadstype ved journalføring for ettersending og papirsøknader
I tilfelle bruker har valgt feil stønadstype må saksbehandler kunne overskrive stønadstypen. Valget i dropdownen settes automatisk til stønadstypen som ligger på journalposten
